fix: storybook、testファイル及びフォルダ構成の見直し Close #158
