+++ components/atoms/nav/stories/nav.stories.js
@@ -2,14 +2,13 @@ import { storiesOf } from '@storybook/vue';
 import { text, boolean } from '@storybook/addon-knobs/vue';
 import NavIcon from '../NavIcon.vue';
 import NavImg from '../NavImg.vue';
-import NavText from '../NavText.vue';
 import NavLogo from '../NavLogo.vue';
 
 storiesOf('Atoms/Nav/NavLogo', module).add(
   'default',
   () => ({
     components: { NavLogo },
-    template: '<nav-logo :nav-logo="text" />',
+    template: '<nav-logo>Text</nav-logo>',
     props: {
       text: {
         type: String,
@@ -58,25 +57,6 @@ storiesOf('Atoms/Nav/NavImg', module).add(
   {
     info: {
       summary: 'ナビゲーションの画像表示',
-    },
-  }
-);
-
-storiesOf('Atoms/Nav/NavText', module).add(
-  'default',
-  () => ({
-    components: { NavText },
-    template: '<nav-text :nav-text="text" />',
-    props: {
-      text: {
-        type: String,
-        default: text('text', 'text'),
-      },
-    },
-  }),
-  {
-    info: {
-      summary: 'ナビゲーションのテキスト',
     },
   }
 );
